Refactor database schema and add triggers and views
This commit refactors the database schema by adding two new tables: "reporte_categoria" and "auditoria_categoria". It also adds a trigger "tg_categoria_creada" that is fired after inserting a new row in the "categoria" table. The trigger logs the creation of a category in the "auditoria_categoria" table. Additionally, a view "reporte_categorias_creadas" is created to display the number of categories created by each admin.

The commit also includes the removal of a redundant comment and the addition of a new stored procedure "sp_agregar_inscripcion_curso" for adding an enrollment to a course.

These changes improve the functionality and maintainability of the database.